+++ src/main/kotlin/no/nav/tiltaksarrangor/model/Endringsmelding.kt
@@ -31,7 +31,7 @@ data class Endringsmelding(
 
 		data class EndreDeltakelseProsentInnhold(
 			val deltakelseProsent: Int,
-			val dagerPerUke: Float?,
+			val dagerPerUke: Int?,
 			val gyldigFraDato: LocalDate?
 		) : Innhold()
